pyproject.toml
[project]
name = "deep-learning-final-project"
version = "0.1.0"
description = "Academic Research Project: Deep Learning Final Project - A structured data science workflow"
readme = "README.md"
authors = [
    { name = "copilot-swe-agent[bot]", email = "198982749+Copilot@users.noreply.github.com" }
]
requires-python = ">=3.12"
# Production dependencies for deep learning research
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "tensorboard>=2.13.0",
]

[project.optional-dependencies]
# Development dependencies for academic research workflow
dev = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipython>=8.12.0",
    "ipykernel>=6.25.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]

# Additional dependencies for specific tasks
viz = [
    "plotly>=5.15.0",
    "bokeh>=3.2.0",
]

nlp = [
    "transformers>=4.30.0",
    "tokenizers>=0.13.0",
]

cv = [
    "opencv-python>=4.8.0",
    "albumentations>=1.3.0",
]

[project.scripts]
# Command-line tools for common tasks
train = "deep_learning_final_project.scripts.train:main"
evaluate = "deep_learning_final_project.scripts.evaluate:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []
# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []
# Exclude common paths
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]
line-length = 100
# Assume Python 3.12
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src/deep_learning_final_project",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
